feat: get list of all videos on a channel
Use the YouTube Data API to obtain a list of all the videos on a channel. The API has quota limits,
so to avoid repeated use, the list is cached in a SQLite database. Python3 has SQLite support in its
standard library, so this doesn't even require additional packages to use.

The next time a list query is run, if the channel is already in the database, the list will be
retrieved from there.

Right now the channel ID is hard coded, and this ID is actually something that's not completely
straight forward to obtain, so we will replace it with something else, perhaps a video ID, from
which the channel ID can then be obtained through an API call.

Since this little utility is single purpose and not likely to be used by anyone else, I've decided
not to bother unit testing.