feat: use argparse to provide commands
The `list` command will either use the YouTube Data API to get a list of all the videos on a given
channel, or it will get them from the local database cache. If the former, the list will be cached
locally to avoid hitting the YouTube API. Will need to implement an argument that can be used to
sync new videos for channels that are being updated.

The `list-channel` command just prints a simple list of all the channels you've interacted with
before, which stores the channel ID, something which is not easily available in the UI.