feat: provide download command
The download command will iterate over all the videos in the database for a given channel ID and use
`yt-dlp` to download each of them.

Once the video has been downloaded, its entry in the database is updated with the path it was
downloaded to. If the `download` command runs again, it will check if that file exists, and if so,
it will be skipped for re-downloading. Just in case `yt-dlp` is configured with the `--no-overwrite`
option, which has the same effect.

It expects the `YT_CH_ARCHIVER_ROOT_PATH` variable to be set. Under here, it will create a directory
with the channel name, which will then have three subdirectories:

* video
* info
* description

The reason is just to make organisation a bit tidier.

For each video, we will ask `yt-dlp` to get the video along with its thumbnail, info and
description. These get stored in the three directories above, respectively. The thumbnail gets
stored along with the video because though `yt-dlp` is supposed to have a feature to enable the
thumbnail to be saved to a different directory, it doesn't work.

Each of these files is saved as <video-id>.<ext>. The default behaviour for `yt-dlp` is to have the
title of the video in the file name, but this can lead to unwieldy filenames, sometimes with special
characters. The titles for the video are stored in the info files, so we can build an index to
navigate the saved videos, which is the next command I intend to implement.

The other noteworthy thing is that `yt-dlp` is configured to use cookies from Firefox, which should
bypass age verification.