feat: provide get-playlists command
Gets all the playlists for a channel and caches them in the database.

If a video is in the playlist which is not found in the main channel, it is assumed to be an
unlisted video. It is also possible for videos on a playlist to relate to other channels. Those
videos are considered "external".