feat: output video list to spreadsheet
Use an `--xls` argument to output the channel list to a spreadsheet with five columns: ID, title,
status, duration and resolution. If a video has not been downloaded, the row will be marked in red.

Since obtaining the duration and resolution of the videos is very slow, the video downloading
process has been updated such that the resolution and duration are recorded and saved in the
database record at the point where the video has been retrieved.