feat: provide more channels commands
A `channel get` and `channel update` commands are added which use the YouTube data API to pull
additional information for the channels, like description and thumbnails.

Also modifies the video list spreadsheet report to break large channels over 1000 rows. This was
because the thumbnails started to be inserted incorrectly after a certain number of rows, which
happened to be 1536. I think this is either some kind of issue in the spreadsheet library or perhaps
in LibreOffice.