feat: download individual videos
Provides a `--video-id` argument to the `videos download` command for downloading individual videos.

The channel and video information is saved in the cache after it's retrieved. If the video has
already been downloaded we don't do anything.

Going through the process has raised an issue. The currently saved channels don't have the right
usernames. This means we're going to have to do admin to update all channels, saved paths in cached
videos, and download paths on the file systems.