feat: update all channels with new info
Loop over all channels and update them with any new information.

This was implemented mainly to correct the data I had at the moment, which incorrectly assigned the
channel title to the username, or incorrect usernames that were assigned manually.

However, it may come in useful as an admin operation.