feat: provide `playlists download` command
For a given channel, download all items on all playlists, or download the items for a specific
playlist. A playlist item is a video, which doesn't necessarily belong to the channel where the
playlist is defined. We consider that video to be 'external'.

If a video is encountered that wasn't previously in the cache, its information will be retrieved
from YouTube.