fix: save thumbnails as image byte array
The `.tobytes()` function of `PilImage` was not intended to save the image as a byte array. It has
now been replaced with the correct way, which saves the image as a PNG.

All the existing image data that's in the database will need to be replaced with this new mechanism.