chore: remove channel thumbnails
It was more hassle than it was worth to try and store these and they took up a huge amount of space.
For some channels, the image formats didn't seem to be recognised easily by PIL, which meant that
other commands that fetched or saved channels were affected.

The thumbnails columns were removed from the schema then the database was 'vacuumed'; it went from
1.2GB to 9.1MB.